Refactoring: moved download of remote config from MissionControl.java to the RemoteConfig.java. Cleanup of enums and interfaces.
