--- library/src/main/java/com/appculture/missioncontrol/ErrorType.java
@@ -1,13 +1,0 @@
-package com.appculture.missioncontrol;
-
-/**
- * Error types enumeration.
- * <p>
- * Created by abozic on 6/29/16.
- */
-public enum ErrorType {
-
-    NO_REMOTE_URL,
-    BAD_RESPONSE_CODE,
-    INVALID_DATA
-}
