+++ library/src/main/java/com/appculture/missioncontrol/configurations/ErrorType.java
@@ -1,0 +1,13 @@
+package com.appculture.missioncontrol.configurations;
+
+/**
+ * Configuration error types.
+ * <p>
+ * Created by abozic on 9/5/16.
+ */
+public enum ErrorType {
+    NO_REMOTE_URL,
+    BAD_RESPONSE_CODE,
+    CONNECTION_ERROR,
+    INVALID_DATA
+}
